show icon column for checkbox and toggle components
