+++ src/Filament/Tables/UserAttributeColumn.php
@@ -2,6 +2,7 @@
 
 namespace Luttje\FilamentUserAttributes\Filament\Tables;
 
+use Filament\Tables\Columns\Column;
 use Filament\Tables\Columns\TextColumn;
 use Illuminate\Database\Eloquent\Model;
 use Luttje\FilamentUserAttributes\Contracts\HasUserAttributesContract;
@@ -12,7 +13,12 @@ class UserAttributeColumn extends TextColumn
     {
         parent::setUp();
 
-        $this->getStateUsing(function (?Model $record) {
+        self::setUpColumn($this);
+    }
+
+    public static function setUpColumn(Column $column): Column
+    {
+        $column->getStateUsing(function (?Model $record) use ($column) {
             /** @var HasUserAttributesContract */
             $record = $record;
 
@@ -23,7 +29,9 @@ class UserAttributeColumn extends TextColumn
                 throw new \Exception("User attributes not available for instance of $class. Did you forget to implement Luttje\FilamentUserAttributes\Contracts\HasUserAttributesContract with the Luttje\FilamentUserAttributes\Traits\HasUserAttributes trait on this model?");
             }
 
-            return $record->user_attributes->{$this->name};
+            return $record->user_attributes->{$column->name};
         });
+
+        return $column;
     }
 }
